+++ src/main/java/com/atlauncher/gui/tabs/instances/InstancesNavigationPanel.java
@@ -27,34 +27,28 @@ import javax.swing.JButton;
 import javax.swing.JComboBox;
 import javax.swing.JPanel;
 
-import com.atlauncher.viewmodel.base.IInstancesTabViewModel;
 import org.mini2Dx.gettext.GetText;
 
-import com.atlauncher.App;
 import com.atlauncher.evnt.listener.RelocalizationListener;
 import com.atlauncher.evnt.manager.RelocalizationManager;
 import com.atlauncher.gui.dialogs.ImportInstanceDialog;
-import com.atlauncher.gui.tabs.InstancesTab;
 import com.atlauncher.utils.sort.InstanceSortingStrategies;
 import com.atlauncher.utils.sort.InstanceSortingStrategy;
+import com.atlauncher.viewmodel.base.IInstancesTabViewModel;
 
 public final class InstancesNavigationPanel extends JPanel implements RelocalizationListener {
-    private final InstancesTab parent;
     private final IInstancesTabViewModel viewModel;
 
     private final JButton importButton = new JButton(GetText.tr("Import"));
     private final InstancesSearchField searchField;
     private final JComboBox<InstanceSortingStrategy> sortingBox = new JComboBox<>(InstanceSortingStrategies.values());
 
-    public InstancesNavigationPanel(final InstancesTab parent,final IInstancesTabViewModel viewModel) {
-        this.parent = parent;
+    public InstancesNavigationPanel(final IInstancesTabViewModel viewModel) {
         this.viewModel = viewModel;
-        this.searchField = new InstancesSearchField(parent,viewModel);
+        this.searchField = new InstancesSearchField(viewModel);
         this.sortingBox.setMaximumSize(new Dimension(190, 23));
 
-        if (App.settings.defaultInstanceSorting != InstanceSortingStrategies.BY_NAME) {
-            this.sortingBox.setSelectedItem(App.settings.defaultInstanceSorting);
-        }
+        this.sortingBox.setSelectedItem(viewModel.getSort());
 
         this.setLayout(new BoxLayout(this, BoxLayout.X_AXIS));
         this.setBorder(BorderFactory.createEmptyBorder(5, 5, 5, 5));
@@ -75,7 +69,7 @@ public final class InstancesNavigationPanel extends JPanel implements Relocaliza
         // item listeners
         this.sortingBox.addItemListener((e) -> {
             if (e.getStateChange() == ItemEvent.SELECTED) {
-                this.viewModel.setSort((InstanceSortingStrategy) e.getItem());
+                this.viewModel.setSort((InstanceSortingStrategies) e.getItem());
             }
         });
     }